TimeFixed

+++ Assets/Scripts/Stats.cs
@@ -4,10 +4,10 @@ using UnityEngine;
 public class Stats : ScriptableObject
 {
     public int maxLife; //Hits que puede recibir
-    public int baseSpeed; //Velocidad base
-    public int speed; //Velocidad
-    public int baseDashSpeed; //Dasheo base
-    public int dashSpeed; //Dasheo
+    public float baseSpeed; //Velocidad base
+    public float speed; //Velocidad
+    public float baseDashSpeed; //Dasheo base
+    public float dashSpeed; //Dasheo
     public int acuracy; //Punteria (0 -> No da, 100 -> Muy precisa) 
     public float rangeRadius; //Quanto duran las balas hasta que despawnean
     public bool isFreezed; //Esta congelado
